Changed  'OnCompiled' callback; fixed 2nd compile being necessary
The OnCompiled function now is registered to the OnHotReload delegate, and not the OnModuleCompilerFinished delegate anymore.

Before, the initial compilation would trigger a 2nd compilation because after the initial compilation the articy classes would not be found yet.

The OnHotReload delegate is called later than the OnModuleCompilerFinished delegate and therefore the new classes have been added to the class hierarchy list at that point, making the 2nd compile unnecessary. Now a 2nd compile only gets triggered if we are importing (with code changes) while we are already compiling (for example because the user used hot reload himself)

- Also removed some unnecessary commented out code