Bugfix: Asset picker wouldn't forward CDO deltas to instances
Before, the asset picker would directly manipulate an ArticyRef. This would result in instances of a blueprint not updating when the CDO ArticyRef changed.

To that end, the Ref had to be updated using the IPropertyHandle system, as it handles various internal logic if used (mark dirty, CDO default delta instance forwarding, undo/redo etc.)

The system could not be used at first because the ArticyRef ID was set to read-only.
Changed to Writeable and in customization overwritten to be disabled/read-only again.

Technically "writeable", therefore IPropertHandle system works now, but still read-only for the end-user.