Fix: ArticyObjectTileView border not updating
The SArticyRefProperty has a TileView to show the currently selected articy object. The TileView's border is colored with the articy object's assigned color.

Bugfix: Instead of statically assigning the border color once it retrieves the color now per callback, effectively allowing updates to the color when the selected object changes