Refactored import cache and import data asset
Import Cache is now a struct rather than a duplicate UObject: much cleaner approach. The cache is now created at the end of the import rather than at the start so that we can be sure the cache was actually valid (i.e. first import would have an empty cache).

EnsureImportFile renamed to EnsureImportDataAsset as the file is the .articyue4 file. It also generates the data asset if it couldn't be found.

When it is generated during import, the import is cancelled because the generation itself causes a full reimport.