Update 1.0.1: Engine plugin fixes
Compiling as engine plugin has stricter requirements than a project plugin. Every editor exposed uproperty or ufunction now has a category, and the name of the import data asset is a cleaned up version of the original name.

As an engine plugin in 4.24, the plugin crashed when creating the import data asset due to 'Maniac Manfred 3.0' having spaces and periods in them. They are now replaced by underscores.