Fixed issue #27 : ExpressoType type matching
implicit conversion of literals in expressoscripts (e.g. "getProp(IsAFloat) < 3") would result in a float expresso type being compared to an int expresso type.

Meaning: "this" being a float type would then compare to the 3 as a float expresso type, even though it's an int type, which leads to UB.

This is only a partial fix because other operators still operate under the "same-type" assumption, but most common operations are fixed