Added ArticyRefWidgetCustomization System
Inspired by the SequenceCustomizations.

Useful to add new buttons to ArticyRefs with specific articy object types, e.g. adding a Print button to all DialogueFragment types.

Step 1: Inherit from IArticyRefWidgetCustomization and from IArticyRefWidgetCustomizationFactory
Step 2: Create a CustomizationStruct inside the overridden Register function that includes a UI extender
Step 3: Override the SupportsType and CreateCustomization function in the Factory
Step 4: Register a Factory creation delegate with the CustomizationManager inside the ArticyEditor module for a specified class.
Cache the resulting Factory to unregister it later on if necessary