Added Articy GV Runtime debugger
The runtime GV debugger is a tool found in the level toolbar to inspect and edit GVs while the game is running in a PIE session.

It generally works as one might expect, and it triggers OnVariableChanged events.

Generated ArticyGVs are no longer instanced due to custom widgets replacing the default. However, the articy GV asset is sort of broken due to details panel customizations being restrictive. Since the default details were broken before it shouldn't be a problem however.