Refactored ArticyRefWidget Customization
Now uses the ArticyObject as a parameter the user can work with instead of an ArticyRef. This makes it more straight forward to use (users could have used ->Get which would give them a clone of the object at editor-time or crash).
Also makes it more compatible with SArticyRefProperties being used outside of details panels, as it required the original FArticyRef* that was worked with via IPropertyHandle, which is only available in details panel