Added custom ArticyRefPin for BP ArticyRef r-values Part 1
Works as expected.
However, a refactoring of ArticyRefCustomization including widgets is required because the "SArticyRefProperty" that is used for the pin widget really is an ID-Widget. The other Ref variables (Clone ID etc.) are available only via Details Customization, which we don't have on a pin.

So SArticyRefProperty will become SArticyIdProperty and SArticyRefProperty will become an Expandable widget