Refactored ArticyRefCustomization to ArticyIdCustomization Part 2
the SArticyRefProperty is now SArticyIdProperty since that is what it really was.
The new ArticyRef widgets still need to be made and supported.

The class filter is now contained within the class picker and also supports an Exact Class filter button. 

Huge bug encountered & fixed: articy asset picker would flicker on zoom level -1 without a fixed size of 300x300. This was due to menu resizing on Tick in SMenuAnchor. A new combo button class preserves the size and only moves the window, which fixes the flickering.