Changed UArticyObject::FindAsset to use caching & editor only
This speeds up asset lookup at editor time dramatically, as before it would always use the asset registry for every lookup.

The static cache is transient and built up during editor usage, packages are cached as an extra so as long as the packages are valid, it will only access the asset registry once to gather packages.