Changed and fixed class filter behavior in articy asset picker
Before, if the user toggled "Exact class" and then went to change the class filter it would only show the class specified via ArticyClassRestriction.
Changed it so Class picker button is disabled if and only if:
ExactClass = true and ExactClassEditable = false, implying ArticyExactClass = true (not set via UI).

Also the class filter is now always initialized with the top level class restriction (before, it would be the type of the currently selected object).